refactor: remove internal API fetch and stabilize mock structure
